% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CVP_ADMMc}
\alias{CVP_ADMMc}
\title{CV (no folds) penalized precision matrix estimation (c++)}
\usage{
CVP_ADMMc(S_train, S_valid, lam, crit_out = "avg", crit_in = "loss",
  tol_out = 1e-04, tol_in = 1e-04, maxit_out = 10000L,
  maxit_in = 10000L, adjmaxit_out = 10000L, start = "warm",
  trace = "progress")
}
\arguments{
\item{S_train}{pxp sample covariance matrix for training data (denominator n).}

\item{S_valid}{pxp sample covariance matrix for validation data (denominator n).}

\item{lam}{positive tuning parameters for elastic net penalty. If a vector of parameters is provided, they should be in increasing order. Defaults to grid of values \code{10^seq(-5, 5, 0.5)}.}

\item{crit_out}{criterion for convergence in outer (blockwise) loop. Criterion \code{avg} will loop until the average absolute parameter change is less than \code{tol_out} times tolerance multiple. Criterion \code{max} will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than \code{tol_out}. Defaults to \code{avg}.}

\item{crit_in}{criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion \code{loss} will loop until the change in the objective for each response after an iteration is less than \code{tol_in}. Criterion \code{avg} will loop until the average absolute change for each response is less than \code{tol_in} times tolerance multiple. Similary, criterion \code{max} will loop until the maximum absolute change is less than \code{tol_in} times tolerance multiple. Defaults to \code{loss}.}

\item{tol_out}{convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.}

\item{tol_in}{convergence tolerance for inner (lasso) loop. Defaults to 1e-4.}

\item{maxit_out}{maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.}

\item{maxit_in}{maximum number of iterations for inner (lasso) loop. Defaults to 1e4.}

\item{adjmaxit_out}{adjusted maximum number of iterations. During cross validation this option allows the user to adjust the maximum number of iterations after the first \code{lam} tuning parameter has converged (for each \code{alpha}). This option is intended to be paired with \code{warm} starts and allows for "one-step" estimators. Defaults to 1e4.}

\item{start}{specify \code{warm} or \code{cold} start for cross validation. Default is \code{warm}.}

\item{trace}{option to display progress of CV. Choose one of \code{progress} to print a progress bar, \code{print} to print completed tuning parameters, or \code{none}.}
}
\value{
cross validation errors (negative validation likelihood)
}
\description{
Cross validation (no folds) function for GLASSO. This function is to be used with CVP_GLASSO.
}
\keyword{internal}
