% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Parallel.R
\name{CVP_GLASSO}
\alias{CVP_GLASSO}
\title{Parallel CV (uses CV_GLASSOc)}
\usage{
CVP_GLASSO(X = NULL, lam = 10^seq(-5, 5, 0.5), diagonal = FALSE,
  crit.out = c("avg", "max"), crit.in = c("loss", "avg", "max"),
  tol.out = 1e-04, tol.in = 1e-04, maxit.out = 10000, maxit.in = 10000,
  adjmaxit.out = NULL, K = 5, start = c("warm", "cold"), cores = 1,
  trace = c("progress", "print", "none"))
}
\arguments{
\item{X}{nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.}

\item{lam}{positive tuning parameters for elastic net penalty. If a vector of parameters is provided, they should be in increasing order. Defaults to grid of values \code{10^seq(-5, 5, 0.5)}.}

\item{diagonal}{option to penalize the diagonal elements of the estimated precision matrix (\eqn{\Omega}). Defaults to \code{FALSE}.}

\item{crit.out}{criterion for convergence in outer (blockwise) loop. Criterion \code{avg} will loop until the average absolute parameter change is less than \code{tol.out} times tolerance multiple. Criterion \code{max} will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than \code{tol.out}. Defaults to \code{avg}.}

\item{crit.in}{criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion \code{loss} will loop until the change in the objective for each response after an iteration is less than \code{tol.in}. Criterion \code{avg} will loop until the average absolute change for each response is less than \code{tol.in} times tolerance multiple. Similary, criterion \code{max} will loop until the maximum absolute change is less than \code{tol.in} times tolerance multiple. Defaults to \code{loss}.}

\item{tol.out}{convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.}

\item{tol.in}{convergence tolerance for inner (lasso) loop. Defaults to 1e-4.}

\item{maxit.out}{maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.}

\item{maxit.in}{maximum number of iterations for inner (lasso) loop. Defaults to 1e4.}

\item{adjmaxit.out}{adjusted maximum number of iterations. During cross validation this option allows the user to adjust the maximum number of iterations after the first \code{lam} tuning parameter has converged (for each \code{alpha}). This option is intended to be paired with \code{warm} starts and allows for 'one-step' estimators. Defaults to NULL.}

\item{K}{specify the number of folds for cross validation.}

\item{start}{specify \code{warm} or \code{cold} start for cross validation. Default is \code{warm}.}

\item{cores}{option to run CV in parallel. Defaults to \code{cores = 1}.}

\item{trace}{option to display progress of CV. Choose one of \code{progress} to print a progress bar, \code{print} to print completed tuning parameters, or \code{none}.}
}
\value{
returns list of returns which includes:
\item{lam}{optimal tuning parameter.}
\item{min.error}{minimum average cross validation error for optimal parameters.}
\item{avg.error}{average cross validation error across all folds.}
\item{cv.error}{cross validation errors (negative validation likelihood).}
}
\description{
Parallel implementation of cross validation.
}
\keyword{internal}
