% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GLASSOc}
\alias{GLASSOc}
\title{Penalized precision matrix estimation (c++)}
\usage{
GLASSOc(S, initSigma, lam, crit_out = "avg", crit_in = "loss",
  tol_out = 1e-04, tol_in = 1e-04, maxit_out = 10000L,
  maxit_in = 10000L)
}
\arguments{
\item{S}{pxp sample covariance matrix (denominator n).}

\item{initSigma}{initialization matrix for estimated covariance matrix Sigma}

\item{lam}{tuning parameter for lasso penalty.}

\item{crit_out}{criterion for convergence in outer (blockwise) loop. Criterion \code{avg} will loop until the average absolute parameter change is less than \code{tol_out} times tolerance multiple. Criterion \code{max} will loop until the maximum change in the estimated Sigma after an iteration over the parameter set is less than \code{tol_out}. Defaults to \code{avg}.}

\item{crit_in}{criterion for convergence in inner (lasso) loop. Criterion for convergence. Criterion \code{loss} will loop until the change in the objective for each response after an iteration is less than \code{tol_in}. Criterion \code{avg} will loop until the average absolute change for each response is less than \code{tol_in} times tolerance multiple. Similary, criterion \code{max} will loop until the maximum absolute change is less than \code{tol_in} times tolerance multiple. Defaults to \code{loss}.}

\item{tol_out}{convergence tolerance for outer (blockwise) loop. Defaults to 1e-4.}

\item{tol_in}{convergence tolerance for inner (lasso) loop. Defaults to 1e-4.}

\item{maxit_out}{maximum number of iterations for outer (blockwise) loop. Defaults to 1e4.}

\item{maxit_in}{maximum number of iterations for inner (lasso) loop. Defaults to 1e4.}
}
\value{
returns list of returns which includes:
\item{Iterations}{number of iterations.}
\item{lam}{optimal tuning parameters.}
\item{Omega}{estimated penalized precision matrix.}
\item{Sigma}{estimated covariance matrix.}
}
\description{
Penalized precision matrix estimation using the graphical lasso (glasso) algorithm
}
\details{
For details on the implementation of 'GLASSOO', see the vignette
\url{https://mgallow.github.io/GLASSOO/}.
}
\references{
\itemize{
\item 
For more information on the graphical lasso algorithm, see: \cr
Friedman, Jerome, Trevor Hastie, and Robert Tibshirani. "Sparse inverse covariance estimation with the graphical lasso." \emph{Biostatistics} 9.3 (2008): 432-441.\cr
\url{http://statweb.stanford.edu/~tibs/ftp/glasso-bio.pdf}
}
}
\author{
Matt Galloway \email{gall0441@umn.edu}
}
\keyword{internal}
