% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{plot.GLASSO}
\alias{plot.GLASSO}
\title{Plot GLASSO object}
\usage{
\method{plot}{GLASSO}(x, type = c("line", "heatmap"), footnote = TRUE, ...)
}
\arguments{
\item{x}{class object GLASSO}

\item{type}{produce either 'heatmap' or 'line' graph}

\item{footnote}{option to print footnote of optimal values. Defaults to TRUE.}

\item{...}{additional arguments.}
}
\description{
Produces a plot for the cross validation errors, if available.
}
