% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lasso.R
\name{LASSO}
\alias{LASSO}
\title{Lasso Regression (c++)}
\usage{
LASSO(X, Y, lam = 0.1, crit = c("loss", "avg", "max"), tol = 1e-04,
  maxit = 10000, ind = matrix(1, ncol(X), ncol(Y)))
}
\arguments{
\item{X}{nxp data matrix.}

\item{Y}{nxr matrix of response values}

\item{lam}{tuning parameter for lasso regularization term. Defaults to \code{lam = 0.1}.}

\item{crit}{criterion for convergence. Criterion \code{loss} will loop until the relative change in the objective for each response after an iteration is less than \code{tol}. Criterion \code{avg} will loop until the average absolute change for each response is less than \code{tol} times tolerance multiple. Similary, criterion \code{max} will loop until the maximum absolute change is less than \code{tol} times tolerance multiple. Defaults to \code{loss}.}

\item{tol}{tolerance for algorithm convergence. Defaults to 1e-4..}

\item{maxit}{maximum iterations. Defaults to 1e4}

\item{ind}{optional matrix specifying which coefficients will be penalized.}
}
\value{
returns list of returns which includes:
\item{Call}{function call.}
\item{Iterations}{number of iterations.}
\item{Loss}{value of the objective function.}
\item{Coefficients}{estimated regression coefficients.}
}
\description{
Computes the coefficient estimates for lasso-penalized linear regression.
}
\details{
For details on the implementation of 'GLASSO', see the vignette
\url{https:#mgallow.github.io/GLASSO/}.
}
\references{
\itemize{
\item 
For more information on the coordinate descent algorithm, see: \cr
Friedman, Jerome, et al. 'Pathwise coordinate optimization.' \emph{The Annals of
Applied Statistics} 1.2 (2007): 302-332.\cr
\url{https://arxiv.org/pdf/0708.1485.pdf}
}
}
\author{
Matt Galloway \email{gall0441@umn.edu}
}
\keyword{internal}
