% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glasso.R
\name{plot.GLASSO}
\alias{plot.GLASSO}
\title{Plot GLASSO object}
\usage{
\method{plot}{GLASSO}(x, type = c("heatmap", "line"), footnote = TRUE, ...)
}
\arguments{
\item{x}{class object GLASSO}

\item{type}{produce either 'heatmap' or 'line' graph}

\item{footnote}{option to print footnote of optimal values. Defaults to TRUE.}

\item{...}{additional arguments.}
}
\description{
Produces a heat plot for the cross validation errors, if available.
}
\examples{
# generate data from a dense matrix
# first compute covariance matrix
S = matrix(0.9, nrow = 5, ncol = 5)
diag(S) = 1

# generate 100 x 5 matrix with rows drawn from iid N_p(0, S)
Z = matrix(rnorm(100*5), nrow = 100, ncol = 5)
out = eigen(S, symmetric = TRUE)
S.sqrt = out$vectors \%*\% diag(out$values^0.5)
S.sqrt = S.sqrt \%*\% t(out$vectors)
X = Z \%*\% S.sqrt

# produce CV heat map for RIDGEsigma
plot(GLASSO(X, lam = 10^seq(-8, 8, 0.01)))
}
